delete route updated
only the creator of the sauce can delete it now

+++ backend/controllers/sauce.js
@@ -41,7 +41,7 @@ exports.modifySauce = (req,res) => {
     sauceSchema.findOne({ _id: req.params.id })
         .then(sauce => {
             if (sauce.userId =! req.auth.userId) {
-                res.status(401).json({ message : "Non autorisé !" })
+                res.status(403).json({ message : "Non autorisé !" })
             } else {
                 sauceSchema.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id  })
                 .then(() => res.status(200).json({ message: "objet modifié" }))
@@ -54,13 +54,17 @@ exports.modifySauce = (req,res) => {
 exports.deleteSauce = (req,res) => {
     sauceSchema.findOne({ _id: req.params.id })
         .then(sauce => {
-            const filename = sauce.imageUrl.split('/images/')[1];
+            if(sauce.userId != req.auth.userId) {
+                res.status(403).json({ message: "Non autorisé !"})
+            } else {
+                const filename = sauce.imageUrl.split('/images/')[1];
             
-            fs.unlink(`images/${filename}`, () => {
-                sauceSchema.deleteOne({ _id: req.params.id })
-                    .then(() => res.status(200).json({ message: "Objet supprimé !"}))
-                    .catch(error => res.status(400).json({ error }))
+                fs.unlink(`images/${filename}`, () => {
+                    sauceSchema.deleteOne({ _id: req.params.id })
+                        .then(() => res.status(200).json({ message: "Objet supprimé !" }))
+                        .catch(error => res.status(400).json({ error }))
             })
+            }
         })
         .catch(error => res.status(500).json({ error }))
 }
